Add an once-and-for-all workaround for Package::Stash::XS + =< 5.8.6
The simple workaround in ac5c00e9 got broken by P::S::XS 0.36. Since at this
point this is more or less a pattern, add a ridiculous-ish workaround forcing
Package::Stash::PP-only on perls under 5.8.7.

While not ideal, the benefits of ::XS are virtually invisible in this use case
and moreover there is work underway to fix ::XS once and for all. Thus the
commit is considered a viable temporary measure to stave off further errors.