drop Package::Stash dependency
Package::Stash refuses to work with package names that are valid, and
can break on earlier perls. Rather than working around this, just work
with the symbol tree manually. Using Package::Stash doesn't make the
code significantly cleaner or easier to understand. Additionally,
Package::Stash brings in a larger dependency tree than is reasonable for
the work it does.